Support for Real numbers

+++ module_1_part_2_example.py
@@ -6,15 +6,15 @@ handled with an error message.
 """
 
 fav_num = input("Input your favorite number: ")
-fav_num = int(fav_num)
+fav_num = float(fav_num)
 
 least_fav_num = input("Input your least favorite number: ")
-least_fav_num = int(least_fav_num)
+least_fav_num = float(least_fav_num)
 
 while least_fav_num == 0:
     print("Error: Please pick a number other than 0")
     least_fav_num = input("Input your least favorite number: ")
-    least_fav_num = int(least_fav_num)
+    least_fav_num = float(least_fav_num)
 
 ratio = fav_num / least_fav_num
 print("The number fav:least fav number ratio is: " + str(ratio))
